drivers: display: Introduce SSD1357
It is almost identical to SSD1351
