+++ drivers/display/CMakeLists.txt
@@ -24,7 +24,7 @@ zephyr_library_sources_ifdef(CONFIG_SSD1327		ssd1327.c)
 zephyr_library_sources_ifdef(CONFIG_SSD16XX		ssd16xx.c)
 zephyr_library_sources_ifdef(CONFIG_SSD1322		ssd1322.c)
 zephyr_library_sources_ifdef(CONFIG_SSD1331		display_ssd1331.c)
-zephyr_library_sources_ifdef(CONFIG_SSD1351		display_ssd1351.c)
+zephyr_library_sources_ifdef(CONFIG_SSD135X		display_ssd135x.c)
 zephyr_library_sources_ifdef(CONFIG_ST75256		display_st75256.c)
 zephyr_library_sources_ifdef(CONFIG_ST7567		display_st7567.c)
 zephyr_library_sources_ifdef(CONFIG_ST7789V		display_st7789v.c)
