+++ drivers/display/Kconfig
@@ -33,7 +33,7 @@ source "drivers/display/Kconfig.ssd1327"
 source "drivers/display/Kconfig.ssd16xx"
 source "drivers/display/Kconfig.ssd1322"
 source "drivers/display/Kconfig.ssd1331"
-source "drivers/display/Kconfig.ssd1351"
+source "drivers/display/Kconfig.ssd135x"
 source "drivers/display/Kconfig.ssd1363"
 source "drivers/display/Kconfig.st75256"
 source "drivers/display/Kconfig.st7567"
