--- drivers/display/Kconfig.ssd1351
@@ -1,43 +1,0 @@
-# SSD1351 display controller configuration options
-# Copyright (c) 2025 MASSDRIVER EI (massdriver.space)
-# SPDX-License-Identifier: Apache-2.0
-
-menuconfig SSD1351
-	bool "SSD1351 display controller driver"
-	default y
-	depends on DT_HAS_SOLOMON_SSD1351_ENABLED
-	select MIPI_DBI
-	help
-	  Enable driver for SSD1351 display controller.
-
-if SSD1351
-
-config SSD1351_DEFAULT_CONTRAST
-	int "SSD1351 default contrast"
-	default 100
-	range 0 255
-	help
-	  SSD1351 default contrast.
-
-config SSD1351_CONTRASTA
-	int "SSD1351 contrast multiplier for seg A"
-	default 145
-	range 0 255
-	help
-	  Contrast multiplier for seg A
-
-config SSD1351_CONTRASTB
-	int "SSD1351 contrast multiplier for seg B"
-	default 80
-	range 0 255
-	help
-	  Contrast multiplier for seg B
-
-config SSD1351_CONTRASTC
-	int "SSD1351 contrast multiplier for seg C"
-	default 125
-	range 0 255
-	help
-	  Contrast multiplier for seg C
-
-endif # SSD1351
