+++ drivers/display/Kconfig.ssd135x
@@ -1,0 +1,58 @@
+# SSD1351 and SSD1357 display controller configuration options
+# Copyright (c) 2025 MASSDRIVER EI (massdriver.space)
+# SPDX-License-Identifier: Apache-2.0
+
+config SSD135X
+	bool "SSD135x display controllers driver"
+	select MIPI_DBI
+	help
+	  Enable driver for SSD135x display controllers.
+
+config SSD1351
+	bool "SSD1351 display controllers driver"
+	default y
+	depends on DT_HAS_SOLOMON_SSD1351_ENABLED
+	select SSD135X
+	help
+	  Enable driver for SSD1351 display controller.
+
+
+config SSD1357
+	bool "SSD1357 display controllers driver"
+	default y
+	depends on DT_HAS_SOLOMON_SSD1357_ENABLED
+	select SSD135X
+	help
+	  Enable driver for SSD1357 display controller.
+
+if SSD135X
+
+config SSD135X_DEFAULT_CONTRAST
+	int "SSD135x default contrast"
+	default 100
+	range 0 255
+	help
+	  Default contrast.
+
+config SSD135X_CONTRASTA
+	int "SSD135x contrast multiplier for seg A"
+	default 145
+	range 0 255
+	help
+	  Contrast multiplier for seg A
+
+config SSD135X_CONTRASTB
+	int "SSD135x contrast multiplier for seg B"
+	default 80
+	range 0 255
+	help
+	  Contrast multiplier for seg B
+
+config SSD135X_CONTRASTC
+	int "SSD135x contrast multiplier for seg C"
+	default 125
+	range 0 255
+	help
+	  Contrast multiplier for seg C
+
+endif # SSD135X
