+++ dts/bindings/display/solomon,ssd1357.yaml
@@ -1,8 +1,8 @@
 # Copyright (c) 2025 MASSDRIVER EI (massdriver.space)
 # SPDX-License-Identifier: Apache-2.0
 
-description: SSD1351 display controller
+description: SSD1357 display controller
 
-compatible: "solomon,ssd1351"
+compatible: "solomon,ssd1357"
 
 include: ["solomon,ssd135x-common.yaml"]
